+++ progettoSicurezzaSV/Test/UserTest.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 void test_user() {
 	const int EXPECTED_ID = 1;
-	const float EXPECTED_THRESHOLD = 0.70;
+	const float EXPECTED_THRESHOLD = (float) 0.70;
 
 	User* to_test = new User(1);
 	int returned_id = to_test->get_id();
@@ -33,7 +33,7 @@ void test_user() {
 
 void test_set_threshold() {
 	const int RANDOM_ID = 1;
-	const float EXPECTED_THRESHOLD = 0.20;
+	const float EXPECTED_THRESHOLD = (float) 0.20;
 
 	User* to_test = new User(1);
 	to_test->set_threshold(EXPECTED_THRESHOLD);
